Improves inventory management and block placement
Refactors the inventory system to manage item stacks correctly when placing blocks.

This change removes unnecessary inventory updates and ensures that item amounts are correctly decremented when blocks are placed. It also adds the player inventory initialization on player creation.

Also renames `SendInventory` and `SendInventorySlot` to `UpdateInventory` and `UpdateInventorySlot` respectively.

+++ Minesharp.Server/Storages/StorageSlot.cs
@@ -6,15 +6,15 @@ public class StorageSlot
 {
     public SlotType Type { get; set; }
     public ItemStack Item { get; set; }
-    public int Index { get; }
+    public short Index { get; }
     public bool IsEmpty => Item == null;
     
-    public StorageSlot(int index) : this(SlotType.Container, index)
+    public StorageSlot(short index) : this(SlotType.Container, index)
     {
         
     }
 
-    public StorageSlot(SlotType type, int index)
+    public StorageSlot(SlotType type, short index)
     {
         Type = type;
         Index = index;
